Separated Controller and Model
