+++ api/functions/index.js
@@ -66,45 +66,14 @@ app.get('/', (req, res) => {
 // sets the biography for the user
 app.post("/api/user/profile/set_biography/:email", (req, res) => {
     (async () => {
-        try {
-
-            // get the document to be changed
-            const profileDoc = fsdb.collection("profile").doc(req.params.email);
-
-            // update the document with the request's body
-            await profileDoc.update(req.body);
-
-            // send a response if it's successful
-            return res.status(200).send();
-        } catch (error) {
-
-            // was not able to add the biography due to an error
-            console.log(error);
-
-            // return a 500 response due to error
-            return res.status(500).send(error);
-        }
+        return await data_access.setProfileSection(fsdb, req, res);
     })();
 });
 
 // set profile picture for the user specified
 app.post("/api/user/profile/set_profilepic/:email", (req, res) => {
     (async () => {
-        try {
-
-            // get the document to be changed
-            const profileDoc = fsdb.collection("profile").doc(req.params.email);
-
-            // update the document with the request's body
-            await profileDoc.update(req.body);
-
-            // send a response if it's successful
-            return res.status(200).send();
-        } catch (error) {
-            // return a 500 error if you can't set profile picture for some reason
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.setProfileSection(fsdb, req, res);
 
     })();
 });
@@ -112,96 +81,28 @@ app.post("/api/user/profile/set_profilepic/:email", (req, res) => {
 // sets the social media for the user
 app.post("/api/user/profile/set_social_media/:email", (req, res) => {
     (async () => {
-        try {
-            // get the document to be changed
-            const profileDoc = fsdb.collection("profile").doc(req.params.email);
-
-            // update the document with the request's body
-            await profileDoc.update(req.body);
-
-            // send a response if it's successful
-            return res.status(200).send();
-        } catch (error) {
-
-            // was not able to add the biography due to an error
-            console.log(error);
-
-            // return a 500 response due to error
-            return res.status(500).send(error);
-        }
+        return await data_access.setProfileSection(fsdb, req, res);
     })();
 });
 
 // add a new friend to the current user's friend list
 app.post("/api/friends/add/:currentEmail/:otherEmail", (req, res) => {
     (async () => {
-        try {
-            
-            // get the current user's user document
-            const userDoc = fsdb.collection("user").doc(req.params.currentEmail);
-            await userDoc.update({
-                friends: admin.firestore.FieldValue.arrayUnion(req.params.otherEmail)
-            });
-
-            return res.status(200).send();
-            // update the user's list by adding the 
-        } catch (error) {
-
-            // was not able to add a friend due to an error
-            console.log(error);
-
-            // return a 500 response due to error
-            return res.status(500).send(error);
-        }
+        return await data_access.addFriend(admin, fsdb, req, res);
     })();
 });
 
 // add a new friend to the current user's friend list
 app.post("/api/friends/remove/:currentEmail/:otherEmail", (req, res) => {
     (async () => {
-        try {
-            
-            // get the current user's user document
-            const userDoc = fsdb.collection("user").doc(req.params.currentEmail);
-            await userDoc.update({
-                friends: admin.firestore.FieldValue.arrayRemove(req.params.otherEmail)
-            });
-
-            return res.status(200).send();
-            // update the user's list by adding the 
-        } catch (error) {
-
-            // was not able to remove friend due to an error
-            console.log(error);
-
-            // return a 500 response due to error
-            return res.status(500).send(error);
-        }
+        return await data_access.removeFriend(admin, fsdb, req, res);
     })();
 });
 
 // Go through every user in our database and compute compatibility score
 app.post("/api/in_harmony/:currUserId/:distanceLimit", (req, res) => {
     (async () => {
-        try {
-            var query = fsdb.collection('user');
-            var reset = {
-                similar_users: []
-            };
-
-            // delete document first
-            await fsdb.collection("in_harmony").doc(req.params.currUserId).set(reset);
-
-            var querySnapshot = await query.get();
-                    
-            await in_harmony.populateLeaderboard(admin, fsdb, req.params.currUserId, req.params.distanceLimit, querySnapshot);
-
-            return res.status(200).send();
-
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.setCompatibility(in_harmony, admin, fsdb, req, res);
     })();
 });
 
@@ -360,154 +261,57 @@ app.get('/refresh_token', function(req, res) {
 // get all users
 app.get("/api/users/all", (req, res) => {
     (async() => {
-        try {
-
-            // get all of the users documents
-            const userDoc = await fsdb.collection('user');
-            const users = await userDoc.get();
-
-            return res.status(200).send(users.data());
-        } catch (error) {
-
-            // unable to receive data from database, log and return error code
-            console.log(error);
-            return res.status(500).send(error);
-
-        }
+        return await data_access.getAllUsers(fsdb, req, res);
     })();
 });
 
 // gets all of user's info 
 app.get("/api/user/:email", (req, res) => {
     (async() => {
-        try {
-
-            // get all of the users documents
-            const userDoc = await fsdb.collection('user').doc(req.params.email);
-            const user = await userDoc.get();
-            const data = user.data()
-
-            return res.status(200).send(data);
-        } catch (error) {
-
-            // unable to receive data from database, log and return error code
-            console.log(error);
-            return res.status(500).send(error);
-
-        }
+        return await data_access.userInfo(fsdb, req, res);
     })();
+    
 });
 
-// get the user's name
+// get the user section request
 app.get("/api/user/:section/:email", (req, res) => {
     (async() => {
-        try {
-            
-            // get the section from the user document from database
-            const userDoc = await fsdb.collection('user').doc(req.params.email);
-            const user = await userDoc.get();
-            const data = user.data()[req.params.section];
-            
-
-            // send the user section data back
-            return res.status(200).send(data);
-        } catch (error) {
-
-            // unable to receive data from database, log and return error code
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.userSection(fsdb, req, res);
     })();
 });
 
 // get all of the user's profile from the database
 app.get('/api/profile/:id', (req, res) => {
     (async() => {
-        try {
-            // try getting the information from the database
-            const document = fsdb.collection('profile').doc(req.params.id);
-            let product = await document.get();
-            let response = product.data();
-
-            // send product data to front end
-            return res.status(200).send(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.userProfile(fsdb, req, res);
     })();
 });
 
 // get the profile information from the database
 app.get('/api/profile/:section/:id', (req, res) => {
     (async() => {
-        try {
-            // try getting the information from the database
-            const document = fsdb.collection('profile').doc(req.params.id);
-            let profile = await document.get();
-            let response = profile.data()[req.params.section];
-
-            // send product data to front end
-            return res.status(200).send(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.userProfileSection(fsdb, req, res);
     })();
 });
 
 // get social media from the of the user
 app.get('/api/social/:site/:id', (req, res) => {
     (async() => {
-        try {
-            // try getting the information from the database
-            const document = fsdb.collection('profile').doc(req.params.id);
-            let product = await document.get();
-            let response = product.data()['social_media'][req.params.site];
-            console.log(response);
-
-            // send product data to front end
-            return res.status(200).send(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.getSocial(fsdb, req, res);
     })();
 });
 
 // get the user's stats
 app.get('/api/user/stats/:id', (req, res) => {
     (async() => {
-        try {
-            // try getting the information from the database
-            const document = fsdb.collection('stats').doc(req.params.id);
-            let profile = await document.get();
-            let response = profile.data();
-
-            // send product data to front end
-            return res.status(200).send(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.getAllUserStats(fsdb, req, res);
     })();
 });
 
 // get individual sections of the stats
 app.get('/api/user/stats/:section/:id', (req, res) => {
     (async() => {
-        try {
-            // try getting the information from the database
-            const document = fsdb.collection('stats').doc(req.params.id);
-            let profile = await document.get();
-            let response = profile.data()[req.params.section];
-
-            // send product data to front end
-            return res.status(200).send(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.getUserStatSection(fsdb, req, res);
     })();
 });
 
@@ -522,44 +326,14 @@ app.options('/api/in_harmony/:id', function(req, res) {
 // Get In-Harmony Data from Firestore 
 app.get("/api/in_harmony/:id", (req, res) => {
     (async() => {
-        try {
-            // try getting the information from the database
-            const document = fsdb.collection('in_harmony').doc(req.params.id);
-            let profile = await document.get();
-            let response = profile.data();
-
-            // send product data to front end
-            return res.status(200).send(response);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.getInHarmonyList(fsdb, req, res);
     })();
 })
 
 // Get top artists between two users
 app.get("/api/in_harmony/compare/similar/artists/:currUser/:otherUser", (req, res) => {
     (async() => {
-        try {
-            // Get top stats of current user
-            const document1 = fsdb.collection('stats').doc(req.params.currUser);
-            let profile1 = await document1.get();
-            let response1 = profile1.data();
-            
-            // Get top stats of other user
-            const document2 = fsdb.collection('stats').doc(req.params.otherUser);
-            let profile2 = await document2.get();
-            let response2 = profile2.data();
-
-            // Find similarities
-            var topSimilar = in_harmony.findTopSimilarArtist( response1.top_artists, response2.top_artists );
-
-            // send product data to front end
-            return res.status(200).send(topSimilar);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.getTopArtistTwoUsers(in_harmony, fsdb, req, res);
     })();
 })
 
@@ -567,43 +341,14 @@ app.get("/api/in_harmony/compare/similar/artists/:currUser/:otherUser", (req, re
 // Get top genres between two users
 app.get("/api/in_harmony/compare/similar/genres/:currUser/:otherUser", (req, res) => {
     (async() => {
-        try {
-            // Get top stats of current user
-            const document1 = fsdb.collection('stats').doc(req.params.currUser);
-            let profile1 = await document1.get();
-            let response1 = profile1.data();
-            
-            // Get top stats of other user
-            const document2 = fsdb.collection('stats').doc(req.params.otherUser);
-            let profile2 = await document2.get();
-            let response2 = profile2.data();
-
-            // Find similarities
-            var topSimilar = in_harmony.findTopSimilarGenres( response1.top_genres, response2.top_genres );
-
-            // send product data to front end
-            return res.status(200).send(topSimilar);
-        } catch (error) {
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.getTopGenresTwoUsers(in_harmony, fsdb, req, res);
     })();
 })
 
 // Get comparison between two users
 app.get("/api/in_harmony/compareFriends/:currUserEmail/:friendUserEmail", (req, res) => {
     (async() => {
-        try {
-            // get the compatibility of 2 friends
-            var comparison = await in_harmony.computeFriendCompatibility(fsdb, req.params.currUserEmail, req.params.friendUserEmail);
-
-            // send compatibility data to front end
-            return res.status(200).send(comparison[req.params.friendUserEmail]);
-        } catch (error) {
-            // if error is caught, send 500 and return error message
-            console.log(error);
-            return res.status(500).send(error);
-        }
+        return await data_access.getComparisonsTwoUsers(in_harmony, fsdb, req, res);
     })();
 });
